rename(MOVIES): 폴더명 DTO에서 변경
